// api/models/TaskPage.ts
/**
 * FastAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { ComponentGroup } from '../models/ComponentGroup';
import { Translations } from '../models/Translations';
import { HttpFile } from '../http/http';

export class TaskPage {
    'label'?: Translations | null;
    'componentGroups': Array<ComponentGroup>;
    /**
    * The number of columns used to display the component groups
    */
    'columns'?: number;

    static readonly discriminator: string | undefined = undefined;

    static readonly mapping: {[index: string]: string} | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "label",
            "baseName": "label",
            "type": "Translations",
            "format": ""
        },
        {
            "name": "componentGroups",
            "baseName": "component_groups",
            "type": "Array<ComponentGroup>",
            "format": ""
        },
        {
            "name": "columns",
            "baseName": "columns",
            "type": "number",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return TaskPage.attributeTypeMap;
    }

    public constructor() {
    }
}